client/requirements.txt
python-engineio==3.2.3
python-socketio==3.1.1
six==1.12.0
